refactor(transformers): rename applyBaseImageUrl
