+++ src/transformers/index.js
@@ -7,10 +7,10 @@ const addURLParams = require('./addURLParams')
 const preventWidows = require('./preventWidows')
 const safeClassNames = require('./safeClassNames')
 const replaceStrings = require('./replaceStrings')
+const applyBaseImageUrl = require('./baseImageURL')
 const posthtmlContent = require('./posthtmlContent')
 const removeUnusedCSS = require('./removeUnusedCSS')
 const removeInlineSizes = require('./removeInlineSizes')
-const applyBaseImageUrl = require('./applyBaseImageUrl')
 const applyExtraAttributes = require('./extraAttributes')
 const removeInlineBgColor = require('./removeInlineBgcolor')
 const removeAttributes = require('./posthtmlRemoveAttributes')
